backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.2.0",
    "psycopg-pool>=3.3.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent", "api", "db"]

[dependency-groups]
dev = [
    "supabase>=2.28.3",
]
